requirements.txt
boto3
requests-aws4auth
opensearch-py
PyPDF2